requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.15.1
flower==2.0.1
humanize==4.9.0
kombu==5.3.7
prometheus-client==0.20.0
prompt-toolkit==3.0.43
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
tornado==6.4
typing-extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
